разбивает на модули
